Reset the display  when a number is pressed directly after an equation is run
When an equation is run, the display updates with the result of that equation. If a user wants to continue that equation, they can input an operator, but if they want to start a new one they just need to press a number and now a new equation will be made to interpret.

Also, there was a bugfix for if a user clicks the = key twice in a row: Before, the program would crash, now it doesn't

index.js
- A boolean called equalWasLastClick is used to check for what to do
 - When true, if a number is pressed the display is cleared for a new equation
 - When true, if = is pressed again nothing happens (Bugfix)