Keyboard input first pass
Create the logic for handling keyboard input for 0-9 and all operator keys. The only issue I'm running into so far is that if the page is freshly loaded, the keyboard inputs won't load, but once at least one key has been pressed on the calculator the all work as expected.

This bug was more of an issue before when I had the event listener attached to the number bad alone; in that case, the number keys would only work after a number pad was pressed. They would not work after operators, when the page was first loaded, nor after the = button was pressed.

I plan to work on . and = (enter as well) next, but that will require moving some logic around.

index.js
- Add an event listener to the keypad container (it holds the number and operator pads) to listen for keydown and work with 0-9, +, -, *, and /