Move dot validity checking logic into a function and small bugfix
Move the dot validity logic into a function instead of in the event listener so it can be called for the . key being pressed.

Also, if an equation was run which had just a . for any of its parameters, it would output NaN. However, I'm going to treat . by itself as 0.0, so the logic for that was also added in.

index.js
- In fillEquationArrays() add logic for changing . to 0.0 to prevent NaN
- Move the dot validity logic from an event listener to dotPressValidityCheck()