Fix the weird enter key issues
A stack overflow answer recommended doing preventDefault() for someone who had a similar issue to mine and I decided to try it. It looks like it is working, but since I don't know the default behavior of the enter key I don't actually know why this has fixed it.

index.js
- Inside the event listener for the keypad, add e.preventDefault() to prevent keys (and specifically, the enter key) from doing their default actions