Move the keyboard listener to the document instead of the keypad
I had the event listener for the keyboard linked to the keypad, but apparently when a page loads there is a "focused" place and that place was not the keypad. I wasn't sure what the nature of the issue was, and therefore I didn't want to chase potential issues through Google, and so I asked GPT; it told me about the focus issue and recommended I just put the event listener on the document rather than on the keypad, which is what most sites do (according to GPT).

index.js
- Move the event listener for the keyboard to the document rather than the keypad html element