My final planned commit; lots of small changes
This is the final commit I plan to make for this project. The ony thing I'm really missing from what Odin recommended is to have a clear screen button; I didn't do this because it would mean more time coding the front end, this wouldn't be an issue but it isn't especially important in the wider context of the project. If I were to implement it, it would be an event listener with the following code
displayVal = "";
display.textContent = displayVal;
lastBtnWasOperator = true;
equalWasLastClick = true;

index.js
- Get rid of the old data at the top of the page; this data was used when I was doing the project the Odin way, but once I decided to do it differently this data was no longer needed
 - I never deleted it because I never thought about it since it was at the very top, and I hadn't been working there for a while
- In two places a variable called "temp" was named something more specific
- Some small comment changes here and there
- div() was using one of the old data values and I didn't catch that until I deleted them and all division was suddenly not working
 - Admittedly, this was actually a pretty big oversight as I had changed the operation functions at one point but missed changing all the logic inside div(), luckily this was a small, non-critical project so that oversight wasn't a major issue